refactor: changer les types Long en int pour les IDs dans RentalEntity, RentalController, RentalRepository et RentalService

+++ src/main/java/fr/jessee/chatop/entity/RentalEntity.java
@@ -11,13 +11,21 @@ import lombok.Data;
 public class RentalEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
+
     private String name;
+
     private Double surface;
+
     private Double price;
+
     private String picture;
+
     private String description;
-    private Long ownerId;
+
+    private int ownerId;
+
     private String createdAt;
+
     private String updatedAt;
 }
